# morphio-io/frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Morphio.io API",
		"description": "This API will blow your mind",
		"version": "1.0.0"
	},
	"paths": {
...
		"/template/get-templates": {
			"get": {
				"tags": ["Template"],
				"summary": "Get Templates",
				"operationId": "list_templates",
				"responses": {
					"200": {
						"description": "Templates retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_List_TemplateOut__"
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
		"/template/save-template": {
			"post": {
				"tags": ["Template"],
				"summary": "Save Template",
				"operationId": "save_template",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TemplateCreate"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Template saved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_TemplateOut_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid template content"
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
		"/template/get-template/{template_id}": {
			"get": {
				"tags": ["Template"],
				"summary": "Get Template",
				"operationId": "get_template",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "template_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"title": "Template Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Template retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_TemplateOut_"
								}
							}
						}
					},
					"404": {
						"description": "Template not found"
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/template/update-template/{template_id}": {
			"put": {
				"tags": ["Template"],
				"summary": "Update Template Route",
				"operationId": "update_template",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "template_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"title": "Template Id"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TemplateUpdate"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Template updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_TemplateOut_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid template content"
					},
					"404": {
						"description": "Template not found or unauthorized"
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/template/delete-template/{template_id}": {
			"delete": {
				"tags": ["Template"],
				"summary": "Delete Template Route",
				"operationId": "delete_template",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "template_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"title": "Template Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Template deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_NoneType_"
								}
							}
						}
					},
					"404": {
						"description": "Template not found or unauthorized"
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/user/profile": {
			"get": {
				"tags": ["User", "User"],
				"summary": "Get User Profile",
				"operationId": "get_user_profile",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/app__schemas__auth_schema__UserOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
		"/user/change-display-name": {
			"post": {
				"tags": ["User", "User"],
				"summary": "Change Display Name",
				"operationId": "change_display_name",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdate"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/app__schemas__auth_schema__UserOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
		"/user/change-email": {
			"post": {
				"tags": ["User", "User"],
				"summary": "Change Email",
				"operationId": "change_email",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdate"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/app__schemas__auth_schema__UserOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
		"/user/usage": {
			"get": {
				"tags": ["User", "User"],
				"summary": "Get Current User Usage",
				"operationId": "get_user_usage",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"ResponseModel_List_TemplateOut__": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/TemplateOut"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "Data"
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[List[TemplateOut]]"
			},
...
			"ResponseModel_NoneType_": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"type": "null",
						"title": "Data"
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[NoneType]"
			},
...
			"ResponseModel_TemplateOut_": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/TemplateOut"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[TemplateOut]"
			},
...
			"TemplateCreate": {
				"properties": {
					"name": {
						"type": "string",
						"maxLength": 100,
						"minLength": 1,
						"title": "Name"
					},
					"template_content": {
						"type": "string",
						"minLength": 1,
						"title": "Template Content"
					},
					"is_default": {
						"type": "boolean",
						"title": "Is Default",
						"default": false
					},
					"user_id": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "User Id"
					}
				},
				"type": "object",
				"required": ["name", "template_content"],
				"title": "TemplateCreate"
			},
			"TemplateOut": {
				"properties": {
					"id": {
						"type": "integer",
						"title": "Id"
					},
					"name": {
						"type": "string",
						"title": "Name"
					},
					"template_content": {
						"type": "string",
						"title": "Template Content"
					},
					"user_id": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "User Id"
					},
					"is_default": {
						"type": "boolean",
						"title": "Is Default"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"title": "Created At"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						],
						"title": "Updated At"
					},
					"usage_count": {
						"type": "integer",
						"title": "Usage Count",
						"default": 0
					},
					"user": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/app__schemas__auth_schema__UserOut"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"required": [
					"id",
					"name",
					"template_content",
					"is_default",
					"created_at"
				],
				"title": "TemplateOut"
			},
			"TemplateUpdate": {
				"properties": {
					"name": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Name"
					},
					"template_content": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Template Content"
					}
				},
				"type": "object",
				"title": "TemplateUpdate"
			},
...
			"UserRole": {
				"type": "string",
				"enum": ["USER", "ADMIN", "POWER", "MODERATOR", "GUEST"],
				"title": "UserRole"
			},
			"UserUpdate": {
				"properties": {
					"email": {
						"anyOf": [
							{
								"type": "string",
								"format": "email"
							},
							{
								"type": "null"
							}
						],
						"title": "Email"
					},
					"display_name": {
						"anyOf": [
							{
								"type": "string",
								"maxLength": 50,
								"minLength": 1
							},
							{
								"type": "null"
							}
						],
						"title": "Display Name"
					},
					"role": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/UserRole"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"title": "UserUpdate"
			},
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			},
			"app__schemas__auth_schema__UserOut": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email",
						"title": "Email"
					},
					"display_name": {
						"type": "string",
						"maxLength": 50,
						"minLength": 1,
						"title": "Display Name"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole",
						"default": "USER"
					},
					"id": {
						"type": "integer",
						"title": "Id"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"title": "Created At"
					},
					"last_login": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						],
						"title": "Last Login"
					},
					"is_active": {
						"type": "boolean",
						"title": "Is Active"
					},
					"content_count": {
						"type": "integer",
						"title": "Content Count",
						"default": 0
					}
				},
				"type": "object",
				"required": ["email", "display_name", "id", "created_at", "is_active"],
				"title": "UserOut"
			}
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"type": "oauth2",
				"flows": {
					"password": {
						"scopes": {},
						"tokenUrl": "/auth/login"
					}
				}
			}
		}
	}
}